corrigindo instancia erro id

+++ src/main/java/com/adauto/tatuaria/resources/ClienteResource.java
@@ -1,8 +1,5 @@
 package com.adauto.tatuaria.resources;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -22,7 +19,7 @@ public class ClienteResource {
 	
 	
 	@RequestMapping(value="/{id}" , method = RequestMethod.GET)
-	public ResponseEntity<?> find(@PathVariable Long id) {
+	public ResponseEntity<?> find(@PathVariable Integer id) {
 		
 		Cliente obj = service.buscar(id);
 		return ResponseEntity.ok().body(obj);
